fetch favs from demo1
